+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ## New in git-machete 3.36.4
 
+- fixed: git-machete no longer crashing when branch name contains a `@` (reported by @brandtdaniels)
 - improved: more GitHub PR description template files are recognized (like `.github/PULL_REQUEST_TEMPLATE.md` or `docs/pull_request_template.md`)
 
 ## New in git-machete 3.36.3